+++ src/main/webapp/@storybook/react.js
@@ -1,8 +1,8 @@
 import { storiesOf as of } from '@connexta/ace/@storybook/react'
 
-import withTheme from '../theme/with-theme'
-import withDrawProvider from '../components/location/with-draw-provider'
-import withMuiPickersProvider from '../with-mui-pickers-provider'
+import withTheme from './with-theme'
+import withDrawProvider from './with-draw-provider'
+import withMuiPickersProvider from './with-mui-pickers-provider'
 
 export const storiesOf = (name, m) => {
   const stories = of(name, m)
