+++ src/main/webapp/@storybook/with-draw-provider.tsx
@@ -1,5 +1,5 @@
 import * as React from 'react'
-const { DrawProvider } = require('../../react-hooks/use-draw-interface')
+const { DrawProvider } = require('../react-hooks/use-draw-interface')
 
 const withDrawProvider = (Story: any) => {
   return (
