+++ src/main/webapp/components/maps/cluster-map.js
