+++ src/main/webapp/components/maps/cluster-map.stories.js
