+++ src/main/webapp/components/maps/draw-menu.tsx
