+++ src/main/webapp/components/maps/effects.js
