+++ src/main/webapp/components/maps/icons/bbox.tsx
