+++ src/main/webapp/components/maps/icons/circle.tsx
