+++ src/main/webapp/components/maps/icons/edit-coordinates.tsx
