+++ src/main/webapp/components/maps/icons/index.tsx
