+++ src/main/webapp/components/maps/icons/point.tsx
