+++ src/main/webapp/components/maps/icons/polygon.tsx
