+++ src/main/webapp/components/maps/index.js
