+++ src/main/webapp/components/maps/with-draw-menu.js
