+++ src/main/webapp/components/sharing/group-add-row.js
