+++ src/main/webapp/components/sharing/group-share-panel.js
