+++ src/main/webapp/components/sharing/sharing-modal.js
