+++ src/main/webapp/components/sharing/sharing.js
