+++ src/main/webapp/components/sharing/user-add-row.js
