+++ src/main/webapp/components/sharing/user-share-panel.js
