+++ src/main/webapp/components/simple-search/index.js
@@ -1,0 +1,2 @@
+import SimpleSearch from './simple-search'
+export default SimpleSearch
