+++ src/main/webapp/components/user-settings/result-form-select.js
@@ -4,7 +4,7 @@ import MenuItem from '@material-ui/core/MenuItem'
 import gql from 'graphql-tag'
 import { getIn } from 'immutable'
 import React, { useEffect } from 'react'
-import OutlinedSelect from '../../input/outlined-select'
+import OutlinedSelect from '../input/outlined-select'
 import { useApolloFallback } from '../../react-hooks'
 import LinearProgress from '@material-ui/core/LinearProgress'
 import ErrorMessage from '../network-retry/inline-retry'
